Update Auto-Slayer to v0.19.2: - No longer disables In-Game AutoSlayer: Melvors auto-slayer takes equipped gear when deciding task, this lets users filter task without resorting to skipping monsters in areas they don't want to fight in. When in-game auto-slayer is enabled, auto-slayer skip will still continue to skip tasks that are user set.
Based on suggestion in issue #84.

-------------
- Checks the target area before checking if slayer item is required:
A slayer item is only required when in a slayer area, but some task monsters aren't in a slayer area. (Blue Dragon, Wandering Bard)
This fixes wandering bard causing a undefined error and invalid requirements for other task.

-------------
- `equipmentSlot` is a number, but `SEMIUtils.currentEquipmentInSlot()` expects a string:
Unsure if this changed between versions, but was causing the script to not get the equipped item in the slot.
This just adds a lookup table built on run time using the constants Melvor provides.

Another fix would have `currentEquipmentInSlot` also work with numbers in addition to strings.